+++ target_includes/android_bridges/Textfield.java
@@ -100,8 +100,8 @@ public class Textfield {
 	 */
 	public void setPosition(float x, float y)
 	{
-		int newX = (int)((parentView.getWidth() / 100) * x);
-		int newY = (int)(parentView.getHeight() / 100 * y);
+		float newX = parentView.getWidth() / 100 * x;
+		float newY = parentView.getHeight() / 100 * y;
 		
 		editText.setX(newX);
 		editText.setY(newY);
